Extract test interpreter
