+++ redis/src/main/scala/zio/redis/package.scala
@@ -17,5 +17,8 @@ package object redis
     with options.SortedSets
     with options.Strings
     with options.Lists
-    with options.Streams
-    with Interpreter
+    with options.Streams {
+
+  type RedisExecutor = Has[RedisExecutor.Service]
+
+}
